chore: refactor image input into small components
